+++ CHANGELOG.md
@@ -4,7 +4,12 @@ All notable changes to the eve-utils project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-<!--## [Unreleased]-->
+## [Unreleased]
+### Added
+- /v1/contracts/public/{region_id}/ URL and typings.
+- /v1/contracts/public/bids/{contract_id}/ URL and typings.
+- /v1/contracts/public/items/{contract_id}/ URL and typings.
+
 ## [7.1.0] - 2020-05-18
 ### Added
 - 'strict' parameter to getSearchUrl.